requirements.txt
flask==3.0.0
opencv-python==4.8.1.78
ultralytics==8.0.220
numpy==1.24.3
pillow==10.1.0
scipy==1.11.4
filterpy==1.4.5
scikit-image==0.22.0
lap==0.4.0